Crate subxt_core

subxt-core

A #[no_std] compatible subset of the functionality provided in the subxt crate. This contains the core logic for encoding and decoding things, but nothing related to networking.

Here’s an overview of the main things exposed here:

blocks: decode and explore block bodies.

constants: access and validate the constant addresses in some metadata.

custom_values: access and validate the custom value addresses in some metadata.

metadata: decode bytes into the metadata used throughout this library.

storage: construct storage request payloads and decode the results you’d get back.

tx: construct and sign transactions (extrinsics).

runtime_api: construct runtime API request payloads and decode the results you’d get back.

events: decode and explore events.

Re-exports

pub extern crate alloc;

pub use config::Config;

pub use error::Error;

pub use metadata::Metadata;

Modules

blocks

Decode and iterate over the extrinsics in block bodies.

client

A couple of client types that we use elsewhere.

config

This module provides a Config type, which is used to define various types that are important in order to speak to a particular chain. SubstrateConfig provides a default set of these types suitable for the default Substrate node implementation, and PolkadotConfig for a Polkadot node.

constants

Access constants from metadata.

custom_values

Access custom values from metadata.

dynamic

This module provides the entry points to create dynamic transactions, storage and constant lookups.

error

The errors that can be emitted in this crate.

events

Decode and work with events.

ext

Re-exports of some of the key external crates.

metadata

A Metadata type, which is used through this crate.

runtime_api

Encode runtime API payloads, decode the associated values returned from them, and validate static runtime API payloads.

storage

Encode storage keys, decode storage values, and validate static storage addresses.

tx

Construct and sign transactions.

utils

Miscellaneous utility helpers.