subxt_core

Module blocks

Decode and iterate over the extrinsics in block bodies.

Use the decode_from function as an entry point to decoding extrinsics, and then have a look at Extrinsics and ExtrinsicDetails to see which methods are available to work with the extrinsics.

Example

extern crate alloc;

use subxt_macro::subxt;
use subxt_core::blocks;
use subxt_core::metadata;
use subxt_core::config::PolkadotConfig;
use alloc::vec;

// If we generate types without `subxt`, we need to point to `::subxt_core`:
#[subxt(
    crate = "::subxt_core",
    runtime_metadata_path = "../artifacts/polkadot_metadata_small.scale",
)]
pub mod polkadot {}

// Some metadata we'd like to use to help us decode extrinsics:
let metadata_bytes = include_bytes!("../../../artifacts/polkadot_metadata_small.scale");
let metadata = metadata::decode_from(&metadata_bytes[..]).unwrap();

// Some extrinsics we'd like to decode:
let ext_bytes = vec![
    hex::decode("1004020000").unwrap(),
    hex::decode("c10184001cbd2d43530a44705ad088af313e18f80b53ef16b36177cd4b77b846f2a5f07c01a27c400241aeafdea1871b32f1f01e92acd272ddfe6b2f8b73b64c606572a530c470a94ef654f7baa5828474754a1fe31b59f91f6bb5c2cd5a07c22d4b8b8387350100000000001448656c6c6f").unwrap(),
    hex::decode("550284001cbd2d43530a44705ad088af313e18f80b53ef16b36177cd4b77b846f2a5f07c0144bb92734447c893ab16d520fae0d455257550efa28ee66bf6dc942cb8b00d5d2799b98bc2865d21812278a9a266acd7352f40742ff11a6ce1f400013961598485010000000400008eaf04151687736326c9fea17e25fc5287613693c912909cb226aa4794f26a481700505a4f7e9f4eb106").unwrap()
];

// Given some chain config and metadata, we know how to decode the bytes.
let exts = blocks::decode_from::<PolkadotConfig>(ext_bytes, metadata).unwrap();

// We'll see 3 extrinsics:
assert_eq!(exts.len(), 3);

// We can iterate over them and decode various details out of them.
for ext in exts.iter() {
    println!("Pallet: {}", ext.pallet_name().unwrap());
    println!("Call:   {}", ext.variant_name().unwrap());
}

Re-exports

pub use crate::error::BlockError;

Structs

ExtrinsicDetails

A single extrinsic in a block.

ExtrinsicMetadataDetails

Details for the given extrinsic plucked from the metadata.

ExtrinsicSignedExtension

A single signed extension

ExtrinsicSignedExtensions

The signed extensions of an extrinsic.

Extrinsics

The body of a block.

FoundExtrinsic

A Static Extrinsic found in a block coupled with it’s details.

Traits

StaticExtrinsic

Trait to uniquely identify the extrinsic’s identity from the runtime metadata.

Functions

decode_from

Instantiate a new Extrinsics object, given a vector containing each extrinsic hash (in the form of bytes) and some metadata that we’ll use to decode them.