subxt_core::blocks

Trait StaticExtrinsic

pub trait StaticExtrinsic: DecodeAsFields {
    const PALLET: &'static str;
    const CALL: &'static str;

    // Provided method
    fn is_extrinsic(pallet: &str, call: &str) -> bool { ... }
}

Trait to uniquely identify the extrinsic’s identity from the runtime metadata.

Generated API structures that represent an extrinsic implement this trait.

The trait is utilized to decode emitted extrinsics from a block, via obtaining the form of the Extrinsic from the metadata.

Required Associated Constants

const PALLET: &'static str

Pallet name.

const CALL: &'static str

Call name.

Provided Methods

fn is_extrinsic(pallet: &str, call: &str) -> bool

Returns true if the given pallet and call names match this extrinsic.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.