subxt_core

Module config

This module provides a Config type, which is used to define various types that are important in order to speak to a particular chain. SubstrateConfig provides a default set of these types suitable for the default Substrate node implementation, and PolkadotConfig for a Polkadot node.

Re-exports

pub use polkadot::PolkadotConfig;

pub use polkadot::PolkadotExtrinsicParams;

pub use polkadot::PolkadotExtrinsicParamsBuilder;

pub use signed_extensions::SignedExtension;

pub use substrate::SubstrateConfig;

pub use substrate::SubstrateExtrinsicParams;

pub use substrate::SubstrateExtrinsicParamsBuilder;

Modules

polkadot

Polkadot specific configuration

signed_extensions

This module contains implementations for common signed extensions, each of which implements SignedExtension, and can be used in conjunction with AnyOf to configure the set of signed extensions which are known about when interacting with a chain.

substrate

Substrate specific configuration

Structs

DefaultExtrinsicParamsBuilder

A builder that outputs the set of super::ExtrinsicParams::Params required for DefaultExtrinsicParams. This may expose methods that aren’t applicable to the current chain; such values will simply be ignored if so.

RefineParamsData

Data that can be used to refine the params of signed extensions.

Traits

BlockHash

Block hashes must conform to a bunch of things to be used in Subxt.

Config

Runtime types.

ExtrinsicParams

This trait allows you to configure the “signed extra” and “additional” parameters that are a part of the transaction payload or the signer payload respectively.

ExtrinsicParamsEncoder

This trait is expected to be implemented for any ExtrinsicParams, and defines how to encode the “additional” and “extra” params. Both functions are optional and will encode nothing by default.

Hasher

This represents the hasher used by a node to hash things like block headers and extrinsics.

Header

This represents the block header type used by a node.

RefineParams

Types implementing RefineParams can be modified to reflect live information from the chain.

Type Aliases

DefaultExtrinsicParams

The default super::ExtrinsicParams implementation understands common signed extensions and how to apply them to a given chain.

ParamsFor

given some Config, this return the other params needed for its ExtrinsicParams.