subxt_core::config

Module polkadot

Polkadot specific configuration

Re-exports

pub use crate::utils::AccountId32;

pub use crate::utils::MultiAddress;

pub use crate::utils::MultiSignature;

Structs

H256

Fixed-size uninterpreted hash type with 32 bytes (256 bits) size.

U256

Little-endian large integer type 256-bit unsigned integer.

Enums

PolkadotConfig

Default set of commonly used types by Polkadot nodes.

Type Aliases

PolkadotExtrinsicParams

A struct representing the signed extra and additional parameters required to construct a transaction for a polkadot node.

PolkadotExtrinsicParamsBuilder

A builder which leads to PolkadotExtrinsicParams being constructed. This is what you provide to methods like sign_and_submit().