subxt_core::config

Module signed_extensions

This module contains implementations for common signed extensions, each of which implements SignedExtension, and can be used in conjunction with AnyOf to configure the set of signed extensions which are known about when interacting with a chain.

Structs

AnyOf

This accepts a tuple of SignedExtensions, and will dynamically make use of whichever ones are actually required for the chain in the correct order, ignoring the rest. This is a sensible default, and allows for a single configuration to work across multiple chains.

ChargeAssetTxPayment

The ChargeAssetTxPayment signed extension.

ChargeAssetTxPaymentParams

Parameters to configure the ChargeAssetTxPayment signed extension.

ChargeTransactionPayment

The ChargeTransactionPayment signed extension.

ChargeTransactionPaymentParams

Parameters to configure the ChargeTransactionPayment signed extension.

CheckGenesis

The CheckGenesis signed extension.

CheckMetadataHash

The CheckMetadataHash signed extension.

CheckMortality

The CheckMortality signed extension.

CheckMortalityParams

Parameters to configure the CheckMortality signed extension.

CheckNonce

The CheckNonce signed extension.

CheckNonceParams

Params for CheckNonce

CheckSpecVersion

The CheckSpecVersion signed extension.

CheckTxVersion

The CheckTxVersion signed extension.

Enums

CheckMetadataHashMode

Is metadata checking enabled or disabled?

Traits

SignedExtension

A single SignedExtension has a unique name, but is otherwise the same as ExtrinsicParams in describing how to encode the extra and additional data.