subxt_core::config::signed_extensions

Trait SignedExtension

pub trait SignedExtension<T: Config>: ExtrinsicParams<T> {
    type Decoded: DecodeAsType;

    // Required method
    fn matches(
        identifier: &str,
        _type_id: u32,
        _types: &PortableRegistry,
    ) -> bool;
}

A single SignedExtension has a unique name, but is otherwise the same as ExtrinsicParams in describing how to encode the extra and additional data.

Required Associated Types

type Decoded: DecodeAsType

The type representing the extra bytes of a signed extension. Decoding from this type should be symmetrical to the respective ExtrinsicParamsEncoder::encode_extra_to() implementation of this signed extension.

Required Methods

fn matches(identifier: &str, _type_id: u32, _types: &PortableRegistry) -> bool

This should return true if the signed extension matches the details given. Often, this will involve just checking that the identifier given matches that of the extension in question.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.