subxt_core::config

Trait Config

pub trait Config:
    Sized
    + Send
    + Sync
    + 'static {
    type Hash: BlockHash;
    type AccountId: Debug + Clone + Encode;
    type Address: Debug + Encode + From<Self::AccountId>;
    type Signature: Debug + Encode;
    type Hasher: Debug + Hasher<Output = Self::Hash>;
    type Header: Debug + Header<Hasher = Self::Hasher> + Sync + Send + DeserializeOwned;
    type ExtrinsicParams: ExtrinsicParams<Self>;
    type AssetId: Debug + Clone + Encode + DecodeAsType + EncodeAsType + Send;
}

Runtime types.

Required Associated Types

type Hash: BlockHash

The output of the Hasher function.

type AccountId: Debug + Clone + Encode

The account ID type.

type Address: Debug + Encode + From<Self::AccountId>

The address type.

type Signature: Debug + Encode

The signature type.

type Hasher: Debug + Hasher<Output = Self::Hash>

The hashing system (algorithm) being used in the runtime (e.g. Blake2).

type Header: Debug + Header<Hasher = Self::Hasher> + Sync + Send + DeserializeOwned

The block header.

type ExtrinsicParams: ExtrinsicParams<Self>

This type defines the extrinsic extra and additional parameters.

type AssetId: Debug + Clone + Encode + DecodeAsType + EncodeAsType + Send

This is used to identify an asset in the ChargeAssetTxPayment signed extension.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.