subxt_core::config

Trait ExtrinsicParams

pub trait ExtrinsicParams<T: Config>:
    ExtrinsicParamsEncoder
    + Sized
    + Send
    + 'static {
    type Params: RefineParams<T>;

    // Required method
    fn new(
        client: &ClientState<T>,
        params: Self::Params,
    ) -> Result<Self, ExtrinsicParamsError>;
}

This trait allows you to configure the “signed extra” and “additional” parameters that are a part of the transaction payload or the signer payload respectively.

Required Associated Types

type Params: RefineParams<T>

These parameters can be provided to the constructor along with some default parameters that subxt understands, in order to help construct your ExtrinsicParams object.

Required Methods

fn new( client: &ClientState<T>, params: Self::Params, ) -> Result<Self, ExtrinsicParamsError>

Construct a new instance of our ExtrinsicParams.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.