subxt_core::config

Trait ExtrinsicParamsEncoder

pub trait ExtrinsicParamsEncoder: 'static {
    // Provided methods
    fn encode_extra_to(&self, _v: &mut Vec<u8>) { ... }
    fn encode_additional_to(&self, _v: &mut Vec<u8>) { ... }
}

This trait is expected to be implemented for any ExtrinsicParams, and defines how to encode the “additional” and “extra” params. Both functions are optional and will encode nothing by default.

Provided Methods

fn encode_extra_to(&self, _v: &mut Vec<u8>)

This is expected to SCALE encode the “signed extra” parameters to some buffer that has been provided. These are the parameters which are sent along with the transaction, as well as taken into account when signing the transaction.

fn encode_additional_to(&self, _v: &mut Vec<u8>)

This is expected to SCALE encode the “additional” parameters to some buffer that has been provided. These parameters are not sent along with the transaction, but are taken into account when signing it, meaning the client and node must agree on their values.