subxt_core::config

Trait RefineParams

pub trait RefineParams<T: Config> {
    // Provided method
    fn refine(&mut self, _data: &RefineParamsData<T>) { ... }
}

Types implementing RefineParams can be modified to reflect live information from the chain.

Provided Methods

fn refine(&mut self, _data: &RefineParamsData<T>)

Refine params to an extrinsic. There is usually some notion of ‘the param is already set/unset’ in types implementing this trait. The refinement should most likely not affect cases where a param is in a ‘is already set by the user’ state.