subxt_core::constants

Module address

Construct addresses to access constants with.

Structs

DefaultAddress

This represents the address of a constant.

Traits

Address

This represents a constant address. Anything implementing this trait can be used to fetch constants.

Functions

dynamic

Construct a new dynamic constant lookup.

Type Aliases

DynamicAddress

The type of address typically used to return dynamic constant values.

StaticAddress

The type of address used by our static codegen.