subxt_core

Module dynamic

This module provides the entry points to create dynamic transactions, storage and constant lookups.

Re-exports

pub use crate::tx::payload::dynamic as tx;

pub use crate::constants::address::dynamic as constant;

pub use crate::storage::address::dynamic as storage;

pub use crate::runtime_api::payload::dynamic as runtime_api_call;

Structs

DecodedValueThunk

This is the result of making a dynamic request to a node. From this, we can return the raw SCALE bytes that we were handed back, or we can complete the decoding of the bytes into a DecodedValue type.

Value

Value holds a representation of some value that has been decoded, as well as some arbitrary context.

Traits

At

This trait allows indexing into Values (and options of Values) using the At::at() function. It’s a little like Rust’s [::std::ops::Index] trait, but adapted so that we can return and work with optionals.

Type Aliases

DecodedValue

A scale_value::Value type endowed with contextual information regarding what type was used to decode each part of it. This implements crate::metadata::DecodeWithMetadata, and is used as a return type for dynamic requests.