subxt_core::dynamic

Type Alias DecodedValue

pub type DecodedValue = Value<u32>;

A scale_value::Value type endowed with contextual information regarding what type was used to decode each part of it. This implements crate::metadata::DecodeWithMetadata, and is used as a return type for dynamic requests.

Aliased Type

pub struct DecodedValue {
    pub value: ValueDef<u32>,
    pub context: u32,
}

Fields
value: ValueDef<u32>

The shape and associated data for this Value

context: u32

Some additional arbitrary context that can be associated with a value.