subxt_core::runtime_api

Module payload

This module contains the trait and types used to represent runtime API calls that can be made.

Structs

DefaultPayload

A runtime API payload containing the generic argument data and interpreting the result of the call as ReturnTy.

Traits

Payload

This represents a runtime API payload that can call into the runtime of node.

Functions

dynamic

Create a new DynamicPayload.

Type Aliases

DynamicPayload

A dynamic runtime API payload.

StaticPayload

A statically generated runtime API payload.