subxt_core::storage

Module address

Construct addresses to access storage entries with.

Structs

DefaultAddress

A concrete storage address. This can be created from static values (ie those generated via the subxt macro) or dynamic values via dynamic.

StaticStorageKey

A storage key for static encoded values. The original value is only present at construction, but can be decoded from the contained bytes.

StorageHashers

A collection of storage hashers paired with the type ids of the types they should hash. Can be created for each storage entry in the metadata via StorageHashers::new().

StorageHashersIter

An iterator over all type ids of the key and the respective hashers. See StorageHashers::iter().

Traits

Address

This represents a storage address. Anything implementing this trait can be used to fetch and iterate over storage entries.

StorageKey

This trait should be implemented by anything that can be used as one or multiple storage keys.

Functions

dynamic

Construct a new dynamic storage lookup.

Type Aliases

DynamicAddress

A typical storage address constructed at runtime rather than via the subxt macro; this has no restriction on what it can be used for (since we don’t statically know).

StaticAddress

A storage address constructed by the static codegen.