subxt_core::storage::address

Type Alias DynamicAddress

pub type DynamicAddress<Keys> = DefaultAddress<Keys, DecodedValueThunk, Yes, Yes, Yes>;

A typical storage address constructed at runtime rather than via the subxt macro; this has no restriction on what it can be used for (since we don’t statically know).

Aliased Type

pub struct DynamicAddress<Keys> { /* private fields */ }

Implementations

impl<Keys: StorageKey> DynamicAddress<Keys>

pub fn new( pallet_name: impl Into<String>, entry_name: impl Into<String>, keys: Keys, ) -> Self

Creates a new dynamic address. As Keys you can use a Vec<scale_value::Value>