subxt_core

Module tx

Construct and sign transactions.

Example

use subxt_signer::sr25519::dev;
use subxt_macro::subxt;
use subxt_core::config::PolkadotConfig;
use subxt_core::config::DefaultExtrinsicParamsBuilder as Params;
use subxt_core::tx;
use subxt_core::utils::H256;
use subxt_core::metadata;

// If we generate types without `subxt`, we need to point to `::subxt_core`:
#[subxt(
    crate = "::subxt_core",
    runtime_metadata_path = "../artifacts/polkadot_metadata_small.scale",
)]
pub mod polkadot {}

// Gather some other information about the chain that we'll need to construct valid extrinsics:
let state = tx::ClientState::<PolkadotConfig> {
    metadata: {
        let metadata_bytes = include_bytes!("../../../artifacts/polkadot_metadata_small.scale");
        metadata::decode_from(&metadata_bytes[..]).unwrap()
    },
    genesis_hash: {
        let h = "91b171bb158e2d3848fa23a9f1c25182fb8e20313b2c1eb49219da7a70ce90c3";
        let bytes = hex::decode(h).unwrap();
        H256::from_slice(&bytes)
    },
    runtime_version: tx::RuntimeVersion {
        spec_version: 9370,
        transaction_version: 20,
    }
};

// Now we can build a balance transfer extrinsic.
let dest = dev::bob().public_key().into();
let call = polkadot::tx().balances().transfer_allow_death(dest, 10_000);
let params = Params::new().tip(1_000).nonce(0).build();

// We can validate that this lines up with the given metadata:
tx::validate(&call, &state.metadata).unwrap();

// We can build a signed transaction:
let signed_call = tx::create_signed(&call, &state, &dev::alice(), params).unwrap();

// And log it:
println!("Tx: 0x{}", hex::encode(signed_call.encoded()));

Re-exports

pub use crate::client::ClientState;

pub use crate::client::RuntimeVersion;

Modules

payload

This module contains the trait and types used to represent transactions that can be submitted.

signer

A library to submit extrinsics to a substrate node via RPC.

Structs

PartialTransaction

This represents a partially constructed transaction that needs signing before it is ready to submit. Use PartialTransaction::signer_payload() to return the payload that needs signing, PartialTransaction::sign() to sign the transaction using a SignerT impl, or PartialTransaction::sign_with_address_and_signature() to apply an existing signature and address to the transaction.

Transaction

This represents a signed transaction that’s ready to be submitted. Use Transaction::encoded() or Transaction::into_encoded() to get the bytes for it, or Transaction::hash() to get the hash.

Functions

call_data

Return the SCALE encoded bytes representing the call data of the transaction.

create_partial_signed

Create a partial extrinsic.

create_signed

Creates a signed extrinsic without submitting it.

create_unsigned

Creates an unsigned extrinsic without submitting it.

validate

Run the validation logic against some extrinsic you’d like to submit. Returns Ok(()) if the call is valid (or if it’s not possible to check since the call has no validation hash). Return an error if the call was not valid or something went wrong trying to validate it (ie the pallet or call in question do not exist at all).