subxt_core::tx

Function create_signed

pub fn create_signed<T, Call, Signer>(
    call: &Call,
    client_state: &ClientState<T>,
    signer: &Signer,
    params: <T::ExtrinsicParams as ExtrinsicParams<T>>::Params,
) -> Result<Transaction<T>, Error>
where
    T: Config,
    Call: Payload,
    Signer: SignerT<T>,

Creates a signed extrinsic without submitting it.

Note: if not provided, the default account nonce will be set to 0 and the default mortality will be immortal. This is because this method runs offline, and so is unable to fetch the data needed for more appropriate values.