subxt_core::tx

Module payload

This module contains the trait and types used to represent transactions that can be submitted.

Structs

DefaultPayload

A transaction payload containing some generic CallData.

ValidationDetails

Details required to validate the shape of a transaction payload against some metadata.

Traits

Payload

This represents a transaction payload that can be submitted to a node.

Functions

dynamic

Construct a transaction at runtime; essentially an alias to DefaultPayload::new() which provides a Composite value for the call data.

Type Aliases

DynamicPayload

The type of a payload typically used for dynamic transaction payloads.

StaticPayload

The payload type used by static codegen.